1session管理模块增加 2修改计划模块及相关代码 3前台bootstrap等代码编写

+++ src/main/java/com/eve/bookmarks/dao/ScheduleRepository.java
@@ -2,11 +2,12 @@ package com.eve.bookmarks.dao;
 
 import com.eve.bookmarks.entitys.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 import java.util.List;
 
 public interface ScheduleRepository extends CrudRepository<Schedule, Long>, JpaRepository<Schedule, Long> {
-    List<Schedule> findAllByUserId(int userId);
+    List<Schedule> findAllByUserId(Long userId);
 }
