+++ src/main/java/com/eve/bookmarks/entitys/BookMarkMongo.java
@@ -8,15 +8,15 @@ import javax.persistence.Id;
 public class BookMarkMongo {
     @Id
     private String id;
-    private int uid;
+    private Long uid;
     private String value;
     private Long version;
 
-    public int getUid() {
+    public Long getUid() {
         return uid;
     }
 
-    public void setUid(int uid) {
+    public void setUid(Long uid) {
         this.uid = uid;
     }
 
@@ -33,7 +33,7 @@ public class BookMarkMongo {
     public BookMarkMongo(String value) {
         this.value = value;
     }
-    public BookMarkMongo(String value,int uid,Long version) {
+    public BookMarkMongo(String value,Long uid,Long version) {
         this.value = value;
         this.uid = uid;
         this.version = version;
