+++ src/main/java/com/eve/bookmarks/service/impl/UserServiceImpl.java
@@ -37,7 +37,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User get(int id) {
+    public User get(Long id) {
         return userRepository.getById(id);
     }
